+++ validator/api/src/main/java/tech/pegasys/teku/validator/api/ValidatorApiChannel.java
@@ -66,6 +66,8 @@ public interface ValidatorApiChannel extends ChannelInterface {
 
   void subscribeToBeaconCommittee(List<CommitteeSubscriptionRequest> requests);
 
+  void subscribeToSyncCommitteeSubnets(List<SyncCommitteeSubnetSubscription> subscriptions);
+
   void subscribeToPersistentSubnets(Set<SubnetSubscription> subnetSubscriptions);
 
   void sendSignedAttestation(Attestation attestation);