+++ CHANGELOG.md
@@ -16,5 +16,6 @@ For information on changes in released versions of Teku, see the [releases page]
 - Newly created databases will now use LevelDB for storage instead of RocksDB. This uses less memory and has proven to be more stable. Existing databases are unaffected and will continue to use RocksDB.
 - Support for automatic fail-over of eth1-endpoints.  Multiple endpoints can be specified with the new `--eth1-endpoints` CLI option. Thanks to Enrico Del Fante.
 - implement POST `/eth/v1/validator/duties/sync/{epoch}` for Altair fork.
+- implement POST `/eth/v1/validator/sync_committee_subscriptions` for Altair fork.
 
 ### Bug Fixes